test: move retry integ test in different file
